pyproject.toml
[project]
name = "qa-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Artem Aslyamov",email = "aaslyamov.arts@gmail.com"}
]
readme = "README.md"
requires-python = "3.11"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "selenium (>=4.31.0,<5.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "faker (>=37.1.0,<38.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "allure-pytest (>=2.14.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
